Implement dynamic fading
The duty resolution has been increased for smoother experience.

The LED now gradually brightens when user touches the sensor until it can not get any brighter.
If the user stops touching the sensor, it will begin fading immediately at the same rate.

The LED now no longer flickers in the beginning.

There have also been made some changes to the logging in order to decrease the number of messages for better debugging experience.