Add support for two touch sensors
It should be noted that I failed to find any other selection of pins
that would work. GPIO15 and GPIO4 were the only two pins that I could
find that would work for the two touch sensors, so these should not be
modified in the future.

I have also removed the software filter as it appears to be unnecessary.